moved chatgpt logic outside main file
