+++ cmd/main.go
@@ -1,12 +1,8 @@
 package main
 
 import (
-    "bytes"
     "context"
-    "encoding/json"
-    "fmt"
     "log"
-    "net/http"
     "os"
     "strings"
 
@@ -19,78 +15,6 @@ import (
     "github.com/joho/godotenv"
     "log/slog"
 )
-
-type ChatGPTRequest struct {
-    Model    string   `json:"model"`
-    Messages []Message `json:"messages"`
-}
-
-type Message struct {
-    Role    string `json:"role"`
-    Content string `json:"content"`
-}
-
-type ChatGPTResponse struct {
-    Choices []struct {
-        Message struct {
-            Content string `json:"content"`
-        } `json:"message"`
-    } `json:"choices"`
-}
-
-func queryChatGPT(token, prompt string) (string, error) {
-    slog.Debug("Sending request to ChatGPT", "prompt", prompt)
-    url := "https://api.openai.com/v1/chat/completions"
-    reqBody := ChatGPTRequest{
-        Model: "gpt-3.5-turbo",
-        Messages: []Message{
-            {
-                Role:    "user",
-                Content: prompt,
-            },
-        },
-    }
-    jsonReq, err := json.Marshal(reqBody)
-    if err != nil {
-        slog.Error("Error marshalling request body", "error", err)
-        return "", err
-    }
-
-    req, err := http.NewRequest("POST", url, bytes.NewBuffer(jsonReq))
-    if err != nil {
-        slog.Error("Error creating new HTTP request", "error", err)
-        return "", err
-    }
-
-    req.Header.Set("Content-Type", "application/json")
-    req.Header.Set("Authorization", "Bearer "+token)
-
-    client := &http.Client{}
-    resp, err := client.Do(req)
-    if err != nil {
-        slog.Error("Error sending request to ChatGPT", "error", err)
-        return "", err
-    }
-    defer resp.Body.Close()
-
-    slog.Debug("Request sent to ChatGPT, awaiting response...")
-
-    var chatGPTResp ChatGPTResponse
-    if err := json.NewDecoder(resp.Body).Decode(&chatGPTResp); err != nil {
-        slog.Error("Error decoding ChatGPT response", "error", err)
-        return "", err
-    }
-
-    slog.Debug("Received response from ChatGPT", "response", chatGPTResp)
-
-    if len(chatGPTResp.Choices) > 0 {
-        slog.Debug("ChatGPT response content", "content", chatGPTResp.Choices[0].Message.Content)
-        return chatGPTResp.Choices[0].Message.Content, nil
-    }
-
-    slog.Error("No response from ChatGPT")
-    return "", fmt.Errorf("no response from ChatGPT")
-}
 
 func main() {
     // Загружаем переменные окружения из файла .env
@@ -219,7 +143,7 @@ func main() {
         if respond {
             question := c.Message().Text
             slog.Debug("Detected a message for ChatGPT, querying", "question", question)
-            chatGPTResponse, err := queryChatGPT(config.ChatGPTToken, question)
+            chatGPTResponse, err := handler.QueryChatGPT(config.ChatGPTToken, question)
             if err != nil {
                 slog.Error("failed to query ChatGPT", "error", err)
                 return err
